Fixed minor bugs

+++ ee4308_turtlebot_project/src/global_planner.py
@@ -30,7 +30,7 @@ class PriorityQueue:
 
 def getPt(idx):
     x = idx % width
-    y = idx / width
+    y = int(idx / width)
     return (x, y)
 
 def getIdx(pt):
